renders skeletons

+++ graveyard-backend/db/migrate/20200320163645_create_corpses.rb
@@ -2,7 +2,7 @@ class CreateCorpses < ActiveRecord::Migration[6.0]
   def change
     create_table :corpses do |t|
       t.string :name
-      t.integer :speed
+      t.float :speed
       t.integer :flowers_needed
       t.belongs_to :grave, null: false, foreign_key: true
